+++ applications/gaia/skycat/tclutil/util/include/Mem_Map.h
@@ -68,7 +68,7 @@ public:
     // Map a file from an open file descriptor <handle>.  This function
     // will lookup the length of the file if it is not given.
     Mem_Map (int handle, 
-	     int length = -1, 
+	     size_t length = -1, 
 	     int prot = PROT_READ, 
 	     int share = MAP_SHARED, 
 	     void *addr = 0, 
@@ -76,7 +76,7 @@ public:
 
     // Map a file specified by <file_name>.
     Mem_Map (const char file_name[], 
-	     int len = -1, 
+	     size_t len = -1, 
 	     int flags = O_RDWR,
 	     int mode = MMAP_DEFAULT_PERMS, 
 	     int prot = PROT_READ, 
@@ -87,14 +87,14 @@ public:
     // Map a file from an open file descriptor <handle>.  This function
     // will lookup the length of the file if it is not given.
     int map (int handle, 
-	     int length = -1, 
+	     size_t length = -1, 
 	     int prot = PROT_READ, 
 	     int share = MAP_SHARED, 
 	     void *addr = 0,
 	     off_t pos = 0);
 
     // Remap the file associated with <handle_>.
-    int map (int length = -1, 
+    int map (size_t length = -1, 
 	     int prot = PROT_READ, 
 	     int share = MAP_SHARED, 
 	     void *addr = 0, 
@@ -102,7 +102,7 @@ public:
 
     // Map a file specified by <file_name>.
     int map (const char file_name[], 
-	     int len = -1, 
+	     size_t len = -1, 
 	     int flags = O_RDWR,
 	     int mode = MMAP_DEFAULT_PERMS, 
 	     int prot = PROT_READ, 
@@ -135,15 +135,15 @@ public:
     size_t size (void) const;
 
     // Unmap the region starting at <base_addr_>.
-    int unmap (int len = -1);
+    int unmap (size_t len = -1);
 
     // Unmap the region starting at <addr_>.
-    int unmap (void *addr, int len);
+    int unmap (void *addr, size_t len);
 
     // Sync <len> bytes of the memory region to the backing store
     // starting at <base_addr_>.  If <len> == -1 then sync the whole
     // region.
-    int sync (int len = -1, int flags = MS_SYNC);
+    int sync (size_t len = -1, int flags = MS_SYNC);
 
     // Sync <len> bytes of the memory region to the backing store
     // starting at <addr_>.
@@ -152,7 +152,7 @@ public:
     // Change the protection of the pages of the mapped region to <prot>
     // starting at <base_addr_> up to <len> bytes.  If <len> == -1 then
     // change protection of all pages in the mapped region.
-    int protect (int len = -1, int prot = PROT_READ);
+    int protect (size_t len = -1, int prot = PROT_READ);
 
     // Change the protection of the pages of the mapped region to <prot>
     // starting at <addr> up to <len> bytes.
@@ -163,7 +163,7 @@ public:
 
 #if 0
     // Hook into the underlying VM system.
-    int advise (int behavior, int len = -1);
+    int advise (int behavior, size_t len = -1);
 #endif
 
     // Return the underlying <handle_>.
@@ -198,7 +198,7 @@ private:
     // This method does the dirty work of actually calling ::mmap to map
     // the file into memory.
     int map_it (int handle, 
-		int len = -1, 
+		size_t len = -1, 
 		int prot = PROT_READ, 
 		int share = MAP_SHARED, 
 		void *addr = 0, 
@@ -222,7 +222,7 @@ Mem_Map::handle (void) const
 //----------------------------------------------------------------------------
 inline int
 Mem_Map::map (int handle, 
-	      int len, 
+	      size_t len, 
 	      int prot, 
 	      int share, 
 	      void *addr, 
@@ -235,7 +235,7 @@ Mem_Map::map (int handle,
 //----------------------------------------------------------------------------
 // Remap the file associated with <this->handle_>.
 inline int
-Mem_Map::map (int len, 
+Mem_Map::map (size_t len, 
 	      int prot, 
 	      int share, 
 	      void *addr, 
@@ -282,7 +282,7 @@ Mem_Map::size (void) const
 //----------------------------------------------------------------------------
 // Unmap the region starting at <this->base_addr_>.
 inline int
-Mem_Map::unmap (int len)
+Mem_Map::unmap (size_t len)
 {
     return ::munmap ((caddr_t)this->base_addr_, len < 0 ? this->length_ : len);
 }
@@ -291,7 +291,7 @@ Mem_Map::unmap (int len)
 //----------------------------------------------------------------------------
 // Unmap the region starting at <addr_>.
 inline int
-Mem_Map::unmap (void *addr, int len)
+Mem_Map::unmap (void *addr, size_t len)
 {
     return ::munmap ((caddr_t)addr, len < 0 ? this->length_ : len);
 }
@@ -302,7 +302,7 @@ Mem_Map::unmap (void *addr, int len)
 // at <this->base_addr_>.  If <len> == -1 then sync the whole mapped
 // region.
 inline int
-Mem_Map::sync (int len, int flags)
+Mem_Map::sync (size_t len, int flags)
 {
     return ::msync ((caddr_t)this->base_addr_, 
 		    len < 0 ? this->length_ : len, flags);
@@ -325,7 +325,7 @@ Mem_Map::sync (void *addr, size_t len, int flags)
 // starting at <this->base_addr_> up to <len> bytes.  If <len> == -1
 // then change protection of all pages in the mapped region.
 inline int 
-Mem_Map::protect (int len, int prot)
+Mem_Map::protect (size_t len, int prot)
 {
     if (len < 0) {
 	len = this->length_;
@@ -348,7 +348,7 @@ Mem_Map::protect(void *addr, size_t len, int prot)
 //----------------------------------------------------------------------------
 // Hook into the underlying VM system.
 inline int
-Mem_Map::advise (int behavior, int len)
+Mem_Map::advise (int behavior, size_t len)
 {
     if (len < 0) {
 	len = this->length_;
